backend_django/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
click==7.1.2
Django==3.1.2
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==8.0.1
PyJWT==1.7.1
pytz==2020.1
regex==2020.10.23
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
